moved friends utils
