+++ src/pages/api/users/profile/friends/interact-request.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { verifyUser } from "../../../../../utils/auth";
 import { deleteFriendRequest, getFriendRequest } from "../../../../../database/utils/friendRequests";
-import { addFriend } from "../../../../../utils/friends";
+import { addFriend } from "../../../../../database/utils/friends";
 import { getUser } from "../../../../../utils/users";
 
 export default verifyUser(async function InteractRequest(req:NextApiRequest, res:NextApiResponse) {
